Pagination complete

+++ controllers/admin.js
@@ -1,5 +1,6 @@
 const Product = require('../models/product');
 const { validationResult } = require('express-validator');
+const fileHelper = require('../util/file');
 
 exports.getAddProduct = (req,res,next) => {
     res.render('admin/edit-product', {
@@ -14,10 +15,24 @@ exports.getAddProduct = (req,res,next) => {
 
 exports.postAddProduct = (req,res,next)=>{
     const title = req.body.title;
-    const imageUrl = req.file;
+    const image = req.file;
     const price = req.body.price;
     const description = req.body.description;
-    console.log(imageUrl);
+    if (!image) {
+        return res.status(422).render('admin/edit-product', {
+          pageTitle: 'Add Product',
+          path: '/admin/add-product',
+          editing: false,
+          hasError: true,
+          product: {
+            title: title,
+            price: price,
+            description: description
+          },
+          errorMessage: 'Attached file is not an image.',
+          validationErrors: []
+        });
+    }
     const errors = validationResult(req);
 
     if(!errors.isEmpty()) {
@@ -36,6 +51,9 @@ exports.postAddProduct = (req,res,next)=>{
             }
         });
     }
+
+    const imageUrl = image.path;
+
     const product = new Product({title: title, price: price, description: description, imageUrl: imageUrl, userId: req.user});
     product.save()
     .then(result => {
@@ -95,7 +113,7 @@ exports.postEditProduct = (req, res, next) => {
     const prodId = req.body.productId;
     const updatedTitle = req.body.title;
     const updatedPrice = req.body.price;
-    const updatedImageUrl = req.body.imageUrl;
+    const image = req.file;
     const updatedDesc = req.body.description;
     const errors = validationResult(req);
     if(!errors.isEmpty()) {
@@ -110,7 +128,6 @@ exports.postEditProduct = (req, res, next) => {
                 title: updatedTitle,
                 price: updatedPrice,
                 description: updatedDesc,
-                imageUrl: updatedImageUrl,
                 _id: prodId
             }
         });
@@ -123,7 +140,10 @@ exports.postEditProduct = (req, res, next) => {
         product.title = updatedTitle;
         product.price = updatedPrice;
         product.description = updatedDesc;
-        product.imageUrl = updatedImageUrl;
+        if (image) {
+            fileHelper.deleteFile(product.imageUrl);
+            product.imageUrl = image.path;
+        }
         return product.save().then(result => {
             console.log('Updated Product');
             res.redirect('/admin/products');
@@ -153,7 +173,14 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req,res,next) => {
     const prodId = req.body.productId;
-    Product.deleteOne({_id: prodId , userId: req.user._id})
+    Product.findById(prodId)
+    .then(product=> {
+        if(!product){
+            return next(new Error('Product not found.'))
+        }
+        fileHelper.deleteFile(product.imageUrl);
+        return Product.deleteOne({_id: prodId , userId: req.user._id});
+    })
     .then(() => {
         console.log('Destroyed Product');
         res.redirect('/admin/products');
